fixes to DBF

+++ Wedge.java
@@ -32,6 +32,6 @@ public class Wedge {
 	}
 	
 	public int getTime(){
-		return time;
+		return (int) time;
 	}
 }
